refactor(ela-cli): aggregate producer and crc commands
